Collect headless related stuff under feature flag
